+++ patzilla/access/depatech/expression.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# (c) 2017 Andreas Motl, Elmyra UG
+# (c) 2017-2018 Andreas Motl <andreas.motl@ip-tools.org>
 import types
 import logging
 import pyparsing
@@ -17,15 +17,18 @@ from patzilla.util.python import _exception_traceback
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
 
+
 class DepaTechGrammar(CQLGrammar):
     def preconfigure(self):
         CQLGrammar.preconfigure(self)
         self.cmp_single = u':'.split()
 
+
 class DepaTechParser(object):
 
     def __init__(self, expression=None, modifiers=None):
         self.expression = expression
+        self.search = None
         self.query_object = None
 
     def parse(self):
@@ -33,9 +36,9 @@ class DepaTechParser(object):
         if self.query_object:
             return self
 
-        # Parse expression, extract and propagate keywords
-        self.query_object, query_recompiled = cql_prepare_query(
-            self.expression, grammar=DepaTechGrammar, keyword_fields=DepaTechExpression.fieldnames)
+        # Parse CQL expression and extract keywords
+        self.search = cql_prepare_query(self.expression, grammar=DepaTechGrammar, keyword_fields=DepaTechExpression.fieldnames)
+        self.query_object = self.search.cql_parser
 
         return self
 
@@ -49,8 +52,12 @@ class DepaTechParser(object):
         self.parse()
         return self.query_object.dumps()
 
+    @property
     def keywords(self):
+
         self.parse()
+
+        # Extract classes from representation like "IC:H04L0012433"
         self.rewrite_classes_ops()
 
         keywords = self.query_object.keywords()
@@ -59,6 +66,11 @@ class DepaTechParser(object):
         keywords = unique_sequence(keywords)
 
         return keywords
+
+    @property
+    def keywords_origin(self):
+        return 'lucene'
+
 
 class DepaTechExpression(object):